+++ src/app/quiz/form/SelectQuestionType.tsx
@@ -2,12 +2,14 @@ import { QuestionType } from '@/types/quizzes';
 
 const SelectQuestionType = ({
   id,
+  currentType,
   defaultChecked,
   onChange,
   type,
   title
 }: {
   id: string | undefined;
+  currentType: QuestionType;
   defaultChecked: boolean;
   onChange: (id: string | undefined, type: QuestionType) => void;
   type: QuestionType;
@@ -18,6 +20,7 @@ const SelectQuestionType = ({
       <input
         type="radio"
         name={id}
+        checked={currentType === type}
         defaultChecked={defaultChecked}
         className="mr-2"
         onChange={() => onChange(id, type)}
